back-end/app/requirements.txt
werkzeug==1.0.0
Flask==1.0.2
Flask-API==1.1
redis==2.10.5
requests==2.22.0
docutils==0.14
python-dateutil==2.7.5
jmespath==0.9.3
urllib3==1.25.3
pytz==2019.3
jaeger-client==4.4.0
opentracing==2.4.0
pycryptodome==3.9.9
python-magic==0.4.22

loglib-python
